fix: 修复部分bug,添加设置备份

+++ AutoBookKeeping/app/src/main/java/com/beta/autobookkeeping/activity/orderItemSearch/items/SearchOrderListViewAdapter.java
@@ -115,7 +115,7 @@ public class SearchOrderListViewAdapter extends BaseAdapter {
                 bundle.putInt("month",cursor.getInt(2));
                 bundle.putInt("day",cursor.getInt(3));
                 bundle.putString("clock",cursor.getString(4));
-                bundle.putFloat("money",cursor.getFloat(5));
+                bundle.putDouble("money",cursor.getDouble(5));
                 bundle.putString("bankName",cursor.getString(6));
                 bundle.putString("orderRemark",cursor.getString(7));
                 bundle.putString("costType",cursor.getString(8));
